move certificate and certificatetype into its own module
