+++ runtime/src/validated_block_finalization.rs
@@ -7,7 +7,7 @@
 use {
     crate::bank::Bank,
     agave_votor_messages::{
-        consensus_message::{Certificate, CertificateType},
+        certificate::{Certificate, CertificateType},
         fraction::Fraction,
     },
     log::warn,
